+++ src/fortune.rs
